third_party/WebKit/LayoutTests/http/tests/eventsource/workers/eventsource-url-attribute-expected.txt

CONSOLE WARNING: 'EventSource.URL' is deprecated. Please use 'EventSource.url' instead.
[Worker] Both .URL and .url should work (for compatibility reasons).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: ../script-tests/eventsource-url-attribute.js
PASS [Worker] source.URL is "http://127.0.0.1:8000/eventsource/resources/event-stream.php"
PASS [Worker] source.url is "http://127.0.0.1:8000/eventsource/resources/event-stream.php"
PASS [Worker] source.URL === source.url is true
PASS successfullyParsed is true

TEST COMPLETE
